+++ assets/javascripts/kitten/components/grid/sticky-container/index.js
@@ -1,5 +1,4 @@
 import React, {
-  Fragment,
   useRef,
   useState,
   useEffect,
@@ -194,9 +193,9 @@ const StickyContainerBase = (
   }
 
   return (
-    <Fragment>
+    <>
       {(stuck || isSticky === 'always') && (
-        <StyledSpacer containerHeight={containerHeight} />
+        <StyledSpacer className="k-Spacer" containerHeight={containerHeight} />
       )}
       <StyledStickyContainer
         ref={currentStickyContainer}
@@ -206,7 +205,7 @@ const StickyContainerBase = (
       >
         {children}
       </StyledStickyContainer>
-    </Fragment>
+    </>
   )
 }