+++ assets/javascripts/kitten/components/navigation/header-nav/components/logged-out.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import { Context } from './context'
+
+export const LoggedOut = ({ children }) => (
+  <Context.Consumer>
+    {({ isLogged }) => (!isLogged ? children : null)}
+  </Context.Consumer>
+)