+++ assets/javascripts/kitten/constants/dropdown-config.js
@@ -1,0 +1,2 @@
+// In ms. As the dropdown content is display after 200ms in css.
+export const DROPDOWN_ANIMATED_DELAY = 210